Faster AXI SREC Bootloader
This bootloader reads srec data stored in the flash memory. 
It reads 256 bytes of memory at once and decodes the SREC elements stored in the buffer one by one.
This bootloader does not use any interrupt, it waits for the SPI transfer to finish by using a timer delay. 
The AXI Quad SPI is configured as follows
- Quad SPI Mode in as full AXI
- Clock 66 MHz
- spi x4